photo_text_to_3d/requirements.txt
pillow
numpy
trimesh
matplotlib
rembg
pyglet-1.5.31
onnxruntime